+++ CHANGELOG.md
@@ -6,13 +6,11 @@ Es ersetzt die **Release Notes** aus den Anfangstagen von jFachwert und orientie
 genauso wie an [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Patch-Versionen nicht extra aufgeführt, sondern in der Minor-Version.
 
-## [Planned]
-- Speichervergleich BigDecimal zu PackedDecimal
-
 ## [Unreleased]
 ### Added
 - Anzeigennamen (display name) in E-Mail-Adressen wird unterstützt 
   ([#32](https://github.com/oboehm/jfachwert/issues/32))
+- Speichervergleich BigDecimal zu PackedDecimal
 
 ### Fixed
 - Caching von Bestellnummer.of(..) korrigiert